requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
fpdf2==2.7.8
python-dotenv==1.0.0
apache-airflow==2.8.1
db-dtypes==1.2.0
numpy==1.26.3
